redesign BostonHousing
